AND-7270 [UI] Moved permission BS to another package
