+++ common/ui/src/main/java/com/tangem/common/ui/bottomsheet/permission/state/GiveTxPermissionBottomSheetConfig.kt
@@ -1,4 +1,4 @@
-package com.tangem.common.ui.bottomsheets.state
+package com.tangem.common.ui.bottomsheet.permission.state
 
 import com.tangem.core.ui.components.bottomsheets.TangemBottomSheetConfigContent
