+++ features/swap/presentation/src/main/java/com/tangem/feature/swap/analytics/SwapEvents.kt
@@ -1,6 +1,6 @@
 package com.tangem.feature.swap.analytics
 
-import com.tangem.common.ui.bottomsheets.state.ApproveType
+import com.tangem.common.ui.bottomsheet.permission.state.ApproveType
 import com.tangem.core.analytics.models.AnalyticsEvent
 import com.tangem.feature.swap.domain.models.domain.SwapProvider
 import com.tangem.feature.swap.domain.models.ui.FeeType
